Updated with DApp image and name
This commit adds an image for the DApp and manages the name in the actual start script.
Running this correctly now requires this [PR](https://github.com/status-im/status-dev-cli/pull/15) to be merged. There are not many
changes in the actual PR, so feel free to go ahead and make the necessary adjustments to node_modules/status-dev-cli/cli.js in order
to get the photoPath and name working correctly.